+++ sdk/openai/azure-ai-openai-assistants/src/main/java/com/azure/ai/openai/assistants/models/AssistantCreationOptions.java
@@ -5,6 +5,7 @@ package com.azure.ai.openai.assistants.models;
 
 import com.azure.core.annotation.Fluent;
 import com.azure.core.annotation.Generated;
+import com.azure.core.util.BinaryData;
 import com.azure.json.JsonReader;
 import com.azure.json.JsonSerializable;
 import com.azure.json.JsonToken;
@@ -50,28 +51,12 @@ public final class AssistantCreationOptions implements JsonSerializable<Assistan
     private List<ToolDefinition> tools;
 
     /*
-     * A list of previously uploaded file IDs to attach to the assistant.
-     */
-    @Generated
-    private List<String> fileIds;
-
-    /*
      * A set of up to 16 key/value pairs that can be attached to an object, used for storing additional information
      * about that object in a structured format. Keys may be up to 64 characters in length and values may be up to 512
      * characters in length.
      */
     @Generated
     private Map<String, String> metadata;
-
-    /**
-     * Creates an instance of AssistantCreationOptions class.
-     *
-     * @param model the model value to set.
-     */
-    @Generated
-    public AssistantCreationOptions(String model) {
-        this.model = model;
-    }
 
     /**
      * Get the model property: The ID of the model to use.
@@ -172,28 +157,6 @@ public final class AssistantCreationOptions implements JsonSerializable<Assistan
     }
 
     /**
-     * Get the fileIds property: A list of previously uploaded file IDs to attach to the assistant.
-     *
-     * @return the fileIds value.
-     */
-    @Generated
-    public List<String> getFileIds() {
-        return this.fileIds;
-    }
-
-    /**
-     * Set the fileIds property: A list of previously uploaded file IDs to attach to the assistant.
-     *
-     * @param fileIds the fileIds value to set.
-     * @return the AssistantCreationOptions object itself.
-     */
-    @Generated
-    public AssistantCreationOptions setFileIds(List<String> fileIds) {
-        this.fileIds = fileIds;
-        return this;
-    }
-
-    /**
      * Get the metadata property: A set of up to 16 key/value pairs that can be attached to an object, used for storing
      * additional information about that object in a structured format. Keys may be up to 64 characters in length and
      * values may be up to 512 characters in length.
@@ -227,11 +190,16 @@ public final class AssistantCreationOptions implements JsonSerializable<Assistan
     public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
         jsonWriter.writeStartObject();
         jsonWriter.writeStringField("model", this.model);
+        jsonWriter.writeNumberField("temperature", this.temperature);
+        jsonWriter.writeNumberField("top_p", this.topP);
         jsonWriter.writeStringField("name", this.name);
         jsonWriter.writeStringField("description", this.description);
         jsonWriter.writeStringField("instructions", this.instructions);
         jsonWriter.writeArrayField("tools", this.tools, (writer, element) -> writer.writeJson(element));
-        jsonWriter.writeArrayField("file_ids", this.fileIds, (writer, element) -> writer.writeString(element));
+        jsonWriter.writeJsonField("tool_resources", this.toolResources);
+        if (this.responseFormat != null) {
+            jsonWriter.writeUntypedField("response_format", this.responseFormat.toObject(Object.class));
+        }
         jsonWriter.writeMapField("metadata", this.metadata, (writer, element) -> writer.writeString(element));
         return jsonWriter.writeEndObject();
     }
@@ -249,17 +217,24 @@ public final class AssistantCreationOptions implements JsonSerializable<Assistan
     public static AssistantCreationOptions fromJson(JsonReader jsonReader) throws IOException {
         return jsonReader.readObject(reader -> {
             String model = null;
+            Double temperature = null;
+            Double topP = null;
             String name = null;
             String description = null;
             String instructions = null;
             List<ToolDefinition> tools = null;
-            List<String> fileIds = null;
+            CreateToolResourcesOptions toolResources = null;
+            BinaryData responseFormat = null;
             Map<String, String> metadata = null;
             while (reader.nextToken() != JsonToken.END_OBJECT) {
                 String fieldName = reader.getFieldName();
                 reader.nextToken();
                 if ("model".equals(fieldName)) {
                     model = reader.getString();
+                } else if ("temperature".equals(fieldName)) {
+                    temperature = reader.getNullable(JsonReader::getDouble);
+                } else if ("top_p".equals(fieldName)) {
+                    topP = reader.getNullable(JsonReader::getDouble);
                 } else if ("name".equals(fieldName)) {
                     name = reader.getString();
                 } else if ("description".equals(fieldName)) {
@@ -268,22 +243,158 @@ public final class AssistantCreationOptions implements JsonSerializable<Assistan
                     instructions = reader.getString();
                 } else if ("tools".equals(fieldName)) {
                     tools = reader.readArray(reader1 -> ToolDefinition.fromJson(reader1));
-                } else if ("file_ids".equals(fieldName)) {
-                    fileIds = reader.readArray(reader1 -> reader1.getString());
+                } else if ("tool_resources".equals(fieldName)) {
+                    toolResources = CreateToolResourcesOptions.fromJson(reader);
+                } else if ("response_format".equals(fieldName)) {
+                    responseFormat
+                        = reader.getNullable(nonNullReader -> BinaryData.fromObject(nonNullReader.readUntyped()));
                 } else if ("metadata".equals(fieldName)) {
                     metadata = reader.readMap(reader1 -> reader1.getString());
                 } else {
                     reader.skipChildren();
                 }
             }
-            AssistantCreationOptions deserializedAssistantCreationOptions = new AssistantCreationOptions(model);
+            AssistantCreationOptions deserializedAssistantCreationOptions
+                = new AssistantCreationOptions(model, temperature, topP);
             deserializedAssistantCreationOptions.name = name;
             deserializedAssistantCreationOptions.description = description;
             deserializedAssistantCreationOptions.instructions = instructions;
             deserializedAssistantCreationOptions.tools = tools;
-            deserializedAssistantCreationOptions.fileIds = fileIds;
+            deserializedAssistantCreationOptions.toolResources = toolResources;
+            deserializedAssistantCreationOptions.responseFormat = responseFormat;
             deserializedAssistantCreationOptions.metadata = metadata;
             return deserializedAssistantCreationOptions;
         });
+    }
+
+    /*
+     * A set of resources that are used by the assistant's tools. The resources are specific to the type of tool. For
+     * example, the `code_interpreter`
+     * tool requires a list of file IDs, while the `file_search` tool requires a list of vector store IDs.
+     */
+    @Generated
+    private CreateToolResourcesOptions toolResources;
+
+    /*
+     * What sampling temperature to use, between 0 and 2. Higher values like 0.8 will make the output more random,
+     * while lower values like 0.2 will make it more focused and deterministic.
+     */
+    @Generated
+    private final Double temperature;
+
+    /*
+     * An alternative to sampling with temperature, called nucleus sampling, where the model considers the results of
+     * the tokens with top_p probability mass.
+     * So 0.1 means only the tokens comprising the top 10% probability mass are considered.
+     *
+     * We generally recommend altering this or temperature but not both.
+     */
+    @Generated
+    private final Double topP;
+
+    /*
+     * The response format of the tool calls used by this assistant.
+     */
+    @Generated
+    private BinaryData responseFormat;
+
+    /**
+     * Creates an instance of AssistantCreationOptions class.
+     *
+     * @param model the model value to set.
+     * @param temperature the temperature value to set.
+     * @param topP the topP value to set.
+     */
+    @Generated
+    public AssistantCreationOptions(String model, Double temperature, Double topP) {
+        this.model = model;
+        this.temperature = temperature;
+        this.topP = topP;
+    }
+
+    /**
+     * Creates an instance of AssistantCreationOptions class.
+     *
+     * @param model the model value to set.
+     */
+    public AssistantCreationOptions(String model) {
+        this.model = model;
+        this.temperature = null;
+        this.topP = null;
+    }
+
+
+    /**
+     * Get the toolResources property: A set of resources that are used by the assistant's tools. The resources are
+     * specific to the type of tool. For example, the `code_interpreter`
+     * tool requires a list of file IDs, while the `file_search` tool requires a list of vector store IDs.
+     *
+     * @return the toolResources value.
+     */
+    @Generated
+    public CreateToolResourcesOptions getToolResources() {
+        return this.toolResources;
+    }
+
+    /**
+     * Set the toolResources property: A set of resources that are used by the assistant's tools. The resources are
+     * specific to the type of tool. For example, the `code_interpreter`
+     * tool requires a list of file IDs, while the `file_search` tool requires a list of vector store IDs.
+     *
+     * @param toolResources the toolResources value to set.
+     * @return the AssistantCreationOptions object itself.
+     */
+    @Generated
+    public AssistantCreationOptions setToolResources(CreateToolResourcesOptions toolResources) {
+        this.toolResources = toolResources;
+        return this;
+    }
+
+    /**
+     * Get the temperature property: What sampling temperature to use, between 0 and 2. Higher values like 0.8 will make
+     * the output more random,
+     * while lower values like 0.2 will make it more focused and deterministic.
+     *
+     * @return the temperature value.
+     */
+    @Generated
+    public Double getTemperature() {
+        return this.temperature;
+    }
+
+    /**
+     * Get the topP property: An alternative to sampling with temperature, called nucleus sampling, where the model
+     * considers the results of the tokens with top_p probability mass.
+     * So 0.1 means only the tokens comprising the top 10% probability mass are considered.
+     *
+     * We generally recommend altering this or temperature but not both.
+     *
+     * @return the topP value.
+     */
+    @Generated
+    public Double getTopP() {
+        return this.topP;
+    }
+
+    /**
+     * Get the responseFormat property: The response format of the tool calls used by this assistant.
+     *
+     * @return the responseFormat value.
+     */
+    @Generated
+    public BinaryData getResponseFormat() {
+        return this.responseFormat;
+    }
+
+    /**
+     * Set the responseFormat property: The response format of the tool calls used by this assistant.
+     *
+     * @param responseFormat the responseFormat value to set.
+     * @return the AssistantCreationOptions object itself.
+     */
+    @Generated
+    public AssistantCreationOptions setResponseFormat(BinaryData responseFormat) {
+        this.responseFormat = responseFormat;
+        return this;
     }
 }