+++ sdk/openai/azure-ai-openai-assistants/src/test/java/com/azure/ai/openai/assistants/FunctionsSyncTests.java
@@ -13,6 +13,7 @@ import com.azure.ai.openai.assistants.models.RunStatus;
 import com.azure.ai.openai.assistants.models.RunStep;
 import com.azure.ai.openai.assistants.models.RunStepToolCallDetails;
 import com.azure.ai.openai.assistants.models.SubmitToolOutputsAction;
+import com.azure.ai.openai.assistants.models.ThreadMessageOptions;
 import com.azure.ai.openai.assistants.models.ThreadRun;
 import com.azure.core.http.HttpClient;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -42,10 +43,11 @@ public class FunctionsSyncTests extends AssistantsClientTestBase {
             // Send first user message
             client.createMessage(
                 assistantThread.getId(),
-                MessageRole.USER,
-                "Assuming both my usually preferred vacation spot and favourite airline carrier, how much would it cost "
-                    + "to fly there in September?"
-            );
+                new ThreadMessageOptions(
+                    MessageRole.USER,
+                    "Assuming both my usually preferred vacation spot and favourite airline carrier, how much would it cost "
+                        + "to fly there in September?"
+                ));
 
             // Create run thread
             ThreadRun run = client.createRun(assistantThread, assistant);