+++ sdk/openai/azure-ai-openai-assistants/src/test/java/com/azure/ai/openai/assistants/RunThreadAsyncTest.java
@@ -8,6 +8,7 @@ import com.azure.ai.openai.assistants.models.MessageRole;
 import com.azure.ai.openai.assistants.models.PageableList;
 import com.azure.ai.openai.assistants.models.RunStatus;
 import com.azure.ai.openai.assistants.models.RunStep;
+import com.azure.ai.openai.assistants.models.ThreadMessageOptions;
 import com.azure.ai.openai.assistants.models.ThreadRun;
 import com.azure.core.http.HttpClient;
 import com.azure.core.http.rest.RequestOptions;
@@ -36,7 +37,7 @@ public class RunThreadAsyncTest extends AssistantsClientTestBase {
         String mathTutorAssistantId = createMathTutorAssistant(client);
         String threadId = createThread(client);
         submitMessageAndRunRunner(message -> {
-            StepVerifier.create(client.createMessage(threadId, MessageRole.USER, message))
+            StepVerifier.create(client.createMessage(threadId, new ThreadMessageOptions(MessageRole.USER, message)))
                     .assertNext(threadMessage -> validateThreadMessage(threadMessage, threadId))
                     .verifyComplete();
 
@@ -96,7 +97,7 @@ public class RunThreadAsyncTest extends AssistantsClientTestBase {
         String mathTutorAssistantId = createMathTutorAssistant(client);
         String threadId = createThread(client);
         submitMessageAndRunRunner(message -> {
-            StepVerifier.create(client.createMessage(threadId, MessageRole.USER, message))
+            StepVerifier.create(client.createMessage(threadId, new ThreadMessageOptions(MessageRole.USER, message)))
                     .assertNext(threadMessage -> validateThreadMessage(threadMessage, threadId))
                     .verifyComplete();